requirements.txt
dj-database-url==0.5.0
Django==3.0.5
django-heroku==0.3.1
gunicorn==20.0.4
mysqlclient==1.4.2
psycopg2==2.7.7
pytz==2018.9
whitenoise==5.0.1
django-axes==5.3.1
django-extensions==2.2.8
plotly==4.7.1